refactor: customer 하위 패키지 세분화
