+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.support.Encryption;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.domain.customer.EncryptedPassword;
-import woowacourse.shoppingcart.domain.customer.PlainPassword;
+import woowacourse.shoppingcart.domain.customer.values.password.EncryptedPassword;
+import woowacourse.shoppingcart.domain.customer.values.password.PlainPassword;
 import woowacourse.shoppingcart.dto.CustomerResponse;
 import woowacourse.shoppingcart.dto.CustomerSignUpRequest;
 import woowacourse.shoppingcart.dto.CustomerUpdatePasswordRequest;
