+++ src/main/java/woowacourse/shoppingcart/domain/customer/Customer.java
@@ -1,6 +1,10 @@
 package woowacourse.shoppingcart.domain.customer;
 
 import java.util.Objects;
+import woowacourse.shoppingcart.domain.customer.values.password.EncryptedPassword;
+import woowacourse.shoppingcart.domain.customer.values.Address;
+import woowacourse.shoppingcart.domain.customer.values.PhoneNumber;
+import woowacourse.shoppingcart.domain.customer.values.Username;
 
 public class Customer {
