+++ src/main/java/woowacourse/shoppingcart/domain/customer/values/Username.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain.customer;
+package woowacourse.shoppingcart.domain.customer.values;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
