+++ src/test/java/woowacourse/shoppingcart/dao/CustomerDaoTest.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
 import woowacourse.shoppingcart.domain.customer.Customer;
-import woowacourse.shoppingcart.domain.customer.EncryptedPassword;
+import woowacourse.shoppingcart.domain.customer.values.password.EncryptedPassword;
 
 @JdbcTest
 @AutoConfigureTestDatabase(replace = Replace.NONE)
