+++ src/test/java/woowacourse/shoppingcart/domain/customer/PlainPasswordTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import woowacourse.shoppingcart.domain.customer.values.password.PlainPassword;
 
 public class PlainPasswordTest {
